requirements.txt
# Core Flask
Flask==2.2.5
Werkzeug==3.1.3

# Database
Flask-SQLAlchemy==3.0.5
SQLAlchemy==1.4.49
PyMySQL==1.1.1

# Authentication
Flask-JWT-Extended==4.4.4
PyJWT==2.10.1

# CORS
flask-cors==6.0.1

# PDF Generation
reportlab

# Environment
python-dotenv==1.1.1
